refactor(cyclonedx): Move model mapping to the dedicated file
